PS2 Controller Driver - Arduino to PC (Linux)
This is a driver which uses an Arduino to translate the inputs from a PS2 controller to serial data, and on the Linux side, translates the inputs to joysticks. When the user plugs a controller into the Arduino, the PC recognizes the controller being plugged in and creates the device node automatically, and destroys it when it's done.